TSP GA Recombination
The function tsp_ga_recombination selects pairs of input permutations and recombine them in pairs.
The function new_per takes two permutations A and B as input and find the path of A(1) to A(2) in the B and puts this way in the output. Then puts the rest of path A in output such that if a city in A(3)~A(n) is visited before, it will not be in the output.